refactor: move test_util.ts
