+++ tests/unit/tls_test.ts
@@ -6,7 +6,7 @@ import {
   assertRejects,
   assertStrictEquals,
   assertThrows,
-} from "./test_util.ts";
+} from "../util/test_util.ts";
 import { BufReader, BufWriter } from "@std/io/mod.ts";
 import { readAll } from "@std/streams/read_all.ts";
 import { writeAll } from "@std/streams/write_all.ts";
