+++ tests/unit/write_file_test.ts
@@ -5,7 +5,7 @@ import {
   assertRejects,
   assertThrows,
   unreachable,
-} from "./test_util.ts";
+} from "../util/test_util.ts";
 
 Deno.test(
   { permissions: { read: true, write: true } },
